+++ tests/test_abs.py
@@ -2,9 +2,9 @@
 import pytest
 
 from myia import dtype as ty
-from myia.abstract import ANYTHING, MyiaTypeError
 from myia.prim.py_implementations import typeof
-from myia.abstract.base import (
+from myia.abstract import (
+    ANYTHING, MyiaTypeError,
     AbstractScalar as _S, AbstractTuple as T, AbstractArray as A,
     AbstractList as L, AbstractClass as C,
     amerge,
