+++ tests/test_grad.py
@@ -5,8 +5,7 @@ import numpy as np
 from types import FunctionType
 from dataclasses import dataclass
 
-from myia.abstract import from_value
-from myia.abstract.base import AbstractJTagged
+from myia.abstract import from_value, AbstractJTagged
 from myia.pipeline import standard_resources, standard_pipeline
 from myia.composite import grad
 from myia.debug.finite_diff import GradTester, NoTestGrad, clean_args
